modified services

+++ src/main/java/com/nullhawk/inventory/services/SupplierServiceImp.java
@@ -1,11 +1,12 @@
 package com.nullhawk.inventory.services;
 
+import com.nullhawk.inventory.enums.UserType;
 import com.nullhawk.inventory.exceptions.SupplierNotFoundException;
+import com.nullhawk.inventory.exceptions.UnauthorizedAccessExcpetion;
 import com.nullhawk.inventory.models.Supplier;
 import com.nullhawk.inventory.repositories.SupplierRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-
 import java.util.List;
 
 @Service
@@ -28,23 +29,34 @@ public class SupplierServiceImp implements SupplierService{
     }
 
     @Override
-    public Supplier createSupplier(Supplier supplier) {
-        return supplierRepository.save(supplier);
+    public Supplier createSupplier(UserType user, Supplier supplier) throws UnauthorizedAccessExcpetion {
+        if(user == UserType.ADMIN || user == UserType.MANAGER){
+            return supplierRepository.save(supplier);
+        }else{
+            throw new UnauthorizedAccessExcpetion("Not authorized to change item name");
+        }
     }
 
     @Override
-    public Supplier updateSupplier(Long id, Supplier supplierDetails) {
-        Supplier supplier = getSupplier(id);
-        supplier.setName(supplierDetails.getName());
-        supplier.setContactInformation(supplierDetails.getContactInformation());
-        supplier.setDescription(supplierDetails.getDescription());
-        return supplierRepository.save(supplier);
+    public Supplier updateSupplier(UserType user, Long id, Supplier supplierDetails) throws UnauthorizedAccessExcpetion {
+        if(user == UserType.ADMIN || user == UserType.MANAGER){
+            Supplier supplier = getSupplier(id);
+            supplier.setName(supplierDetails.getName());
+            supplier.setContactInfo(supplierDetails.getContactInfo());
+            supplier.setDescription(supplierDetails.getDescription());
+            return supplierRepository.save(supplier);
+        }else{
+            throw new UnauthorizedAccessExcpetion("Not authorized to change item name");
+        }
     }
 
     @Override
-    public void deleteSupplier(Long id) {
-        Supplier supplier = getSupplier(id);
-        supplierRepository.delete(supplier);
+    public void deleteSupplier(UserType user, Long id) throws UnauthorizedAccessExcpetion {
+        if(user == UserType.ADMIN || user == UserType.MANAGER){
+            Supplier supplier = getSupplier(id);
+            supplierRepository.delete(supplier);
+        }else{
+            throw new UnauthorizedAccessExcpetion("Not authorized to change item name");
+        }
     }
-
 }
