+++ src/main/java/com/nullhawk/inventory/services/TransactionServiceImp.java
@@ -1,6 +1,8 @@
 package com.nullhawk.inventory.services;
 
+import com.nullhawk.inventory.enums.UserType;
 import com.nullhawk.inventory.exceptions.TransactionException;
+import com.nullhawk.inventory.exceptions.UnauthorizedAccessExcpetion;
 import com.nullhawk.inventory.models.Transaction;
 import com.nullhawk.inventory.repositories.TransactionRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -28,23 +30,35 @@ public class TransactionServiceImp implements TransactionService{
     }
 
     @Override
-    public Transaction createTransaction(Transaction transaction) {
-        return transactionRepository.save(transaction);
+    public Transaction createTransaction(UserType user,Transaction transaction) throws UnauthorizedAccessExcpetion {
+        if(user == UserType.ADMIN || user == UserType.MANAGER){
+            return transactionRepository.save(transaction);
+        }else{
+            throw new UnauthorizedAccessExcpetion("Not authorized to change item name");
+        }
     }
 
     @Override
-    public Transaction updateTransaction(Long id, Transaction transactionDetails) {
-        Transaction transaction = getTransaction(id);
-        transaction.setTransactionId(transactionDetails.getTransactionId());
-        transaction.setTransactionType(transactionDetails.getTransactionType());
-        transaction.setQuantity(transactionDetails.getQuantity());
-        transaction.setTransactionDate(transactionDetails.getTransactionDate());
-        return transactionRepository.save(transaction);
+    public Transaction updateTransaction(UserType user,Long id, Transaction transactionDetails) throws UnauthorizedAccessExcpetion {
+        if(user == UserType.ADMIN || user == UserType.MANAGER){
+            Transaction transaction = getTransaction(id);
+            transaction.setId(transactionDetails.getId());
+            transaction.setTransactionType(transactionDetails.getTransactionType());
+            transaction.setAmount(transactionDetails.getAmount());
+            transaction.setDate(transactionDetails.getDate());
+            return transactionRepository.save(transaction);
+        }else{
+            throw new UnauthorizedAccessExcpetion("Not authorized to change item name");
+        } 
     }
 
     @Override
-    public void deleteTransaction(Long id) {
-        Transaction transaction = getTransaction(id);
-        transactionRepository.delete(transaction);
+    public void deleteTransaction(UserType user, Long id) throws UnauthorizedAccessExcpetion {
+        if(user == UserType.ADMIN || user == UserType.MANAGER){
+            Transaction transaction = getTransaction(id);
+            transactionRepository.delete(transaction);
+        }else{
+            throw new UnauthorizedAccessExcpetion("Not authorized to change item name");
+        }
     }
 }
